dotenv_config (dotenv_config v1.1.0)

Types

config_item_name/0

-type config_item_name() :: binary().

config_item_raw_value/0

-type config_item_raw_value() :: binary().

config_item_type/0

-type config_item_type() ::
    simple_config_type() | [exact_value() | simple_config_type()] | convert_function().

config_item_value/0

-type config_item_value() :: any().

convert_function/0

-type convert_function() :: fun((config_item_raw_value()) -> config_item_value()).

exact_value/0

-type exact_value() :: {exact, binary()}.

parsed_config/0

-type parsed_config() :: [{config_item_name(), config_item_value()}].

parser/0

-type parser() :: [{config_item_name(), config_item_type()}].

simple_config_type/0

-type simple_config_type() :: str | int | bool | json | atom | module | charlist.

Functions

get(ConfigItemName)

-spec get(config_item_name()) -> {ok, config_item_value()} | {error, not_found}.

get/1 Get configuration item from the persistent term storage.

get_all()

-spec get_all() -> parsed_config().

get_all/0 Get all configuration items from the persistent term storage.

init(Module, FileNames)

-spec init(module(), [file:name_all()]) -> ok.

init/2 Load configuration from environment variables and override it by provided .env files (if duplicate last provided file wins) and store it in the persistent term storage. Parameters: Module - module which will be used to parse the configuration (see readme for details) FileNames - list of names of the files to load configuration from Returns: ok - if configuration was successfully loaded and stored {error, Reason} - if there was an error during loading or storing

set(ConfigItemName, ConfigItemValue)

-spec set(config_item_name(), config_item_value()) -> ok.

set/2 Set configuration item in the persistent term storage.